Reorganize files in apub lib
