+++ crates/apub/src/activities/comment/create_or_update.rs
@@ -17,11 +17,10 @@ use crate::{
 use activitystreams::{base::AnyBase, link::Mention, primitives::OneOrMany, unparsed::Unparsed};
 use lemmy_api_common::blocking;
 use lemmy_apub_lib::{
+  data::Data,
+  traits::{ActivityFields, ActivityHandler},
   values::PublicUrl,
-  verify_domains_match,
-  ActivityFields,
-  ActivityHandler,
-  Data,
+  verify::verify_domains_match,
 };
 use lemmy_db_queries::Crud;
 use lemmy_db_schema::source::{comment::Comment, community::Community, person::Person, post::Post};
