+++ crates/apub/src/objects/community.rs
@@ -16,7 +16,7 @@ use chrono::{DateTime, FixedOffset};
 use lemmy_api_common::blocking;
 use lemmy_apub_lib::{
   values::{MediaTypeHtml, MediaTypeMarkdown},
-  verify_domains_match,
+  verify::verify_domains_match,
 };
 use lemmy_db_queries::{source::community::Community_, DbPool};
 use lemmy_db_schema::{
