+++ crates/apub/src/objects/person.rs
@@ -15,7 +15,7 @@ use activitystreams::{
 use lemmy_api_common::blocking;
 use lemmy_apub_lib::{
   values::{MediaTypeHtml, MediaTypeMarkdown},
-  verify_domains_match,
+  verify::verify_domains_match,
 };
 use lemmy_db_queries::{source::person::Person_, DbPool};
 use lemmy_db_schema::{
