+++ crates/apub/src/objects/private_message.rs
@@ -14,7 +14,7 @@ use chrono::{DateTime, FixedOffset};
 use lemmy_api_common::blocking;
 use lemmy_apub_lib::{
   values::{MediaTypeHtml, MediaTypeMarkdown},
-  verify_domains_match,
+  verify::verify_domains_match,
 };
 use lemmy_db_queries::{source::private_message::PrivateMessage_, Crud, DbPool};
 use lemmy_db_schema::source::{
