+++ crates/apub_lib/src/verify.rs
@@ -1,0 +1,17 @@
+use activitystreams::error::DomainError;
+use lemmy_utils::LemmyError;
+use url::Url;
+
+pub fn verify_domains_match(a: &Url, b: &Url) -> Result<(), LemmyError> {
+  if a.domain() != b.domain() {
+    return Err(DomainError.into());
+  }
+  Ok(())
+}
+
+pub fn verify_urls_match(a: &Url, b: &Url) -> Result<(), LemmyError> {
+  if a != b {
+    return Err(DomainError.into());
+  }
+  Ok(())
+}
