+++ crates/db_schema/src/source/private_message.rs
@@ -1,7 +1,7 @@
 use crate::{schema::private_message, DbUrl, PersonId, PrivateMessageId};
 use chrono::NaiveDateTime;
 use diesel::{ExpressionMethods, PgConnection, QueryDsl, RunQueryDsl};
-use lemmy_apub_lib::ApubObject;
+use lemmy_apub_lib::traits::ApubObject;
 use lemmy_utils::LemmyError;
 use serde::Serialize;
 use url::Url;
